% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DE_analys}
\alias{DE_analys}
\title{Differential Expression Analysis based on a bottom up model}
\usage{
DE_analys(steps, burnin, rezult, cutoff = 0.3, rope_limits = 0.5)
}
\arguments{
\item{steps}{An integer representing number of iterations of the one step gibbs sampler.}

\item{burnin}{An integer for the number of iterations to be considered as burn in values}

\item{rezult}{A list parameter values at each step}

\item{cutoff}{A threshold value that determines the significance of the difference.}
}
\value{
A logival vector highlighting significantly expressed transcripts.
}
\description{
To determine all the differentially expressed genes across the two conditions
 in a given dataset.
}
