% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DEA}
\alias{DEA}
\title{Differential Expression Analysis based on a bottom up model}
\usage{
DEA(steps, burnin, rezult, cutoff)
}
\arguments{
\item{steps}{An integer representing number of iterations of the gibbs sampler.}

\item{burnin}{An integer for the number of iterations to be considered as burn in values}

\item{rezult}{A list parameter values at each step}

\item{cutoff}{A threshold value that determines the significance of the difference.}
}
\value{
A logival vector highlighting significantly expressed transcripts.
}
\description{
To determine all the differentially expressed genes across the two conditions
 in a given dataset.
}
