% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DEstat}
\alias{DEstat}
\title{Calculate the test statistic.}
\usage{
DEstat(N_A, N_B, rope_limit = 0.5)
}
\arguments{
\item{N_A}{A vector of samples of N_i for condition A.}

\item{N_B}{A vector of samples of N_i for condition B.}

\item{rope-limit}{A float that delimits the range of the region of practical
equivalance, ROPE. A default value of 0.5 is used.}
}
\value{
A value between 0 and 1 that can be interpreted as the probability that there is no difference.
}
\description{
Given two samples of a parameter N_i from each of the two conditions,
 it calculates the area of the distribution of the log2 transformed differences that falls within the region of practical
 equivalance (ROPE) defaulted to [-0.5,0.5].
}
