% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DEtest}
\alias{DEtest}
\title{Differential Expression test}
\usage{
DEtest(N_A, N_B, cutoff)
}
\arguments{
\item{N_A}{A vector of samples of N_i for condition A.}

\item{N_B}{A vector of samples of N_i for condition B.}

\item{cutoff}{A threshold value that determines the significance of the difference.}
}
\value{
A logical value about the significance of difference.
}
\description{
Given two samples of a parameter N_i representing each of the two conditions,
 it determines significance of the differences between the two conditionss based on
 a cutoff value.
}
