% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ercc.R
\docType{data}
\name{ERCC}
\alias{ERCC}
\title{ERCC data.}
\format{A matrix with 71 rows and 10 columns:
\describe{
  \item{Conditions}{There are 5 columns for each of the conditions A and B.}
  \item{Transcripts}{There are 71 distinct transcripts distinguishable by their names.}
}}
\source{
\url{https://bitbucket.org/soccin/seqc/src/ccd0502ef25422e83b3f208f50f8e252f62f17a3/data/?at=master}
}
\usage{
ERCC
}
\description{
A dataset containing a mixture of spike-in synthetic oligonucleotides that
  are mixed into samples A and B at four mixing ratios: 1/2, 2/3, 1 and 4. It has
  been modified from the original dataset by extracting the zero containing  rows.
}
\keyword{datasets}
