% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{get_samples_of}
\alias{get_samples_of}
\title{Get sampled values  of a parameter}
\usage{
get_samples_of(parm, rezult, steps = length(rezult$samples), condition = 1)
}
\arguments{
\item{parm}{A parameter name string i.e  p, f or transcript name.}

\item{rezult}{A list parameter values at each step}

\item{steps}{An integer representing number of iterations of the one step gibbs sampler.}

\item{condition}{An integer representing the two experimental conditions; typically, 1 or 2.}
}
\value{
A vector of parameter samples
}
\description{
To extract samples of individual parameters from the list output of nstepgibbs
produced by  multiple runs of the gibbs sampler.
}
