% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_step_sampler.R
\name{nstepgibbs}
\alias{nstepgibbs}
\title{To estimate model parameters for PCR sequencing using Bayesian Inference.}
\usage{
nstepgibbs(counts, n_A = ncol(counts)/2, n_B = ncol(counts) - n_A,
  m = nrow(counts), counts_A = counts[, 1:n_A], counts_B = counts[, (n_A +
  1):ncol(counts)], steps, tuning.steps = floor(steps/3),
  init.vals = list(N_A = rep(1, m), N_B = rep(1, m), p = 0.005, f = 0.1),
  step.size = list(ss_A = rep(1, m), ss_B = rep(1, m), ss_p = 5e+05, ss_f =
  (1000)))
}
\arguments{
\item{counts}{An m by (n+k) matrix of counts from two experimental conditions}

\item{n_A}{The number of columns of counts in condition A.}

\item{n_B}{The number of columns of counts in condition B.}

\item{counts_A}{An m by n matrix of counts from condition A.}

\item{counts_B}{An m by k matrix of counts from condition B.}

\item{steps}{An integer representing the number of iterations.}

\item{tuning.steps}{An integer representing the number of iterations to be
used for tuning the step sizes.}

\item{init.vals}{A list containing initial values for all the N's (for both
conditions), p and f.}

\item{step.size}{A list containing step sizes values for all the N's, p and
f.}
}
\value{
A list of samples  for each of  \code{N_i} ,\code{p} and \code{f}.
}
\description{
Uses the Gibb's sampling algorithm to sample from the joint distribution of the model parameters.
}
\details{
The counts in each column are used to estimate the size factors which are
in turn used to normalise the counts in each column. The function then uses the
rows of the matrix to estimate parameter N_i for each transcript and uses the
entire dataset (combined from both conditions)  to estimate \code{p} and
\code{f}. The result is an estimate for each of the \code{m} \code{N_i}
parameters for each condition and estimates for \code{p} and \code{f}.
}
