% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\docType{methods}
\name{setInitValues,readsData-method}
\alias{setInitValues,readsData-method}
\title{Mutator method for the initValues slot of the readsData object.}
\usage{
\S4method{setInitValues}{readsData}(object, initval)
}
\arguments{
\item{object}{a readsData object}

\item{initval}{A list of initial values for all the  parameters.}
}
\value{
The same readsData object with the initValues slot updated.
}
\description{
Alters the value of the initValues slot.
}
\examples{
RD <- new("readsData",counts = ERCC)
initvals <- list(N_A = rep(2, 92),N_B = rep(1.5,92), p = 0.0005, f = 0.03)
RD <- setInitValues(RD,initvals)
RD@initValues
}
