% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\name{setInitValues}
\alias{setInitValues}
\title{Mutator method for the initValues slot of the readsData object.}
\usage{
setInitValues(theObject, initval)
}
\arguments{
\item{theObject}{a readsData object}

\item{initval}{A list of initial values for each parameter.}
}
\value{
The same readsData object with the initValues slot updated.
}
\description{
Alters the value of the initValues slot.
}
