% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\docType{methods}
\name{setOutput,readsData-method}
\alias{setOutput,readsData-method}
\title{Mutator method for the output slot of the readsData object.}
\usage{
\S4method{setOutput}{readsData}(object, outval)
}
\arguments{
\item{object}{a readsData object.}

\item{outval}{A list of the output from Bayesian inference.}
}
\value{
The same readsData object with the output slot updated.
}
\description{
sets the value of the output slot.
}
