% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\name{setOutput}
\alias{setOutput}
\title{Mutator method for the output slot of the readsData object.}
\usage{
setOutput(theObject, outval)
}
\arguments{
\item{theObject}{a readsData object.}

\item{outval}{A list of the output from bayesian inference.}
}
\value{
The same readsData object with the output slot updated.
}
\description{
sets the value of the outout slot.
}
