% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\name{setReplicates}
\alias{setReplicates}
\title{Mutator method for the replicates slot of the readsData object.}
\usage{
setReplicates(theObject, repsval)
}
\arguments{
\item{theObject}{a readsData object.}

\item{repsval}{A list of column indices for the samples in each condition.}
}
\value{
The same readsData object with the replicates slot updated.
}
\description{
Alters the value of the replicates slot.
}
