% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\docType{methods}
\name{setStepSizes,readsData-method}
\alias{setStepSizes,readsData-method}
\title{Mutator method for the stepSizes slot of the readsData object.}
\usage{
\S4method{setStepSizes}{readsData}(object, stepSizesval)
}
\arguments{
\item{object}{a readsData object.}

\item{stepSizesval}{A list of step sizes for all the  parameters.}
}
\value{
The same readsData object with the stepSizes slot updated.
}
\description{
Alters the value of the stepSizes slot.
}
\examples{
RD <- new("readsData",counts = ERCC)
ss <- list(N_A = rep(2, 92),N_B = rep(1.5,92), p = 3e5, f = 3.5e7)
RD <- setStepSizes(RD,ss)
RD@stepSizes
}
