% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\name{setStepSizes}
\alias{setStepSizes}
\title{Mutator method for the stepSizes slot of the readsData object.}
\usage{
setStepSizes(theObject, stepSizesval)
}
\arguments{
\item{theObject}{a readsData object.}

\item{stepSizesval}{A list of step sizes for each parameter.}
}
\value{
The same readsData object with the stepSizes slot updated.
}
\description{
Alters the value of the stepSizes slot.
}
