% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{simulated data.}
\format{A matrix with 749 rows and 10 columns:
\describe{
  \item{Conditions}{There are 5 columns for each of the conditions A and B.}
  \item{Transcripts}{There are 749 distinct transcripts without  names.}
}}
\usage{
simdat
}
\description{
A dataset containing simulated data based on parameter values N = 1,2,...,50, p = 0.0017
 and f = 0.5. The values of N were repeated 15 times to generate roughly 750 genomic events.
}
\keyword{datasets}
