% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{tunedStepSize}
\alias{tunedStepSize}
\title{Get tuned values of the step size.}
\usage{
tunedStepSize(RD)
}
\arguments{
\item{RD}{A readsData object with a filled output slot.}
}
\value{
A list of the tuned step sizes of each of the parameter.
}
\description{
To extract samples of individual parameters contianed in the  output slot
produced by  denoiSeq.
}
