% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ercc.R
\docType{data}
\name{ERCC}
\alias{ERCC}
\title{ERCC dataset}
\format{A matrix with 92 rows and 10 columns:
\describe{
  \item{Conditions}{There are 5 columns for each of the conditions A and B.}
  \item{Transcripts}{There are 92 distinct transcripts distinguishable by
  their names.}
}}
\source{
\url{https://bitbucket.org/soccin/seqc/src/ccd0502ef25422e83b3f208f50f8e252f62f17a3/data/?at=master}
}
\usage{
ERCC
}
\description{
RNA-seq data from biological replicates of 3 cell lines. This dataset
 contains a mixture of spike-in synthetic oligonucleotides that
  are mixed into samples A and B at four mixing ratios: 1/2, 2/3, 1 and 4.
}
\keyword{datasets}
