% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{results}
\alias{results}
\title{Extract results from denoiseq analysis}
\usage{
results(RDobject, steps, burnin = floor(steps/3), rope_limit = 0.5)
}
\arguments{
\item{RDobject}{A readsData object with the output slot filled.}

\item{steps}{An integer representing the number of iterations.}

\item{burnin}{An integer for the number of iterations to be considered as burn in values}

\item{rope-limit}{A float that delimits the range of the region of practical
equivalance, ROPE. A default value of 0.5 is used.}
}
\value{
A dataframe with 3 columns namely; the log2 fold change (log2FC), the standard error of the log2 fold change (lgfcSE) and the test static (stat).
}
\description{
To extract the results of Bayesian inference returned by denoiSeq function and
 computes the summary and test statistics.
}
\examples{
#pre -filtering to remove lowly expressed genes
ERCC <- ERCC[rowSums(ERCC)>10,]
RD <- new('readsData',counts = ERCC)
steps <- 100
#100 steps are not adequate. Just for demonstration here.
BI <- denoiseq(RD,steps)
rez <- results(BI,steps)
head(rez)
}
