% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\name{setInitValues}
\alias{setInitValues}
\title{Generic  for the initValues slot of the readsData object.}
\usage{
setInitValues(object, initval)
}
\arguments{
\item{object}{a readsData object}

\item{initval}{A list of initial values for all the  parameters.}
}
\description{
Generic  for the initValues slot of the readsData object.
}
\keyword{internal}
