% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\docType{methods}
\name{setInitValues}
\alias{setInitValues}
\alias{setInitValues,readsData-method}
\title{Generic  for altering the  initValues slot}
\usage{
setInitValues(object, initval)

\S4method{setInitValues}{readsData}(object, initval)
}
\arguments{
\item{object}{a readsData object}

\item{initval}{A list of initial values for each of the  parameters.}
}
\value{
The same readsData object with the initValues slot updated.
}
\description{
Updates the value of the initValues slot for the readsData object supplied.
}
\section{Methods (by class)}{
\itemize{
\item \code{readsData}: Alters the value of the initValues slot of
a readsData object.
}}

\examples{
RD <- new("readsData", counts = ERCC)
initvals <- list(N_A = rep(2, 92), N_B = rep(1.5, 92), p = 0.0005, f = 0.03)
RD <- setInitValues(RD, initvals)
RD@initValues
}
