% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\docType{methods}
\name{setReplicates}
\alias{setReplicates}
\alias{setReplicates,readsData-method}
\title{Generic  for the altering setReplicates slot.}
\usage{
setReplicates(object, repsval)

\S4method{setReplicates}{readsData}(object, repsval)
}
\arguments{
\item{object}{a readsData object}

\item{repsval}{A list  of column indices for the samples in each condition.}
}
\value{
The same readsData object with the replicates slot updated.
}
\description{
Updates the value of the replicates slot for the readsData object supplied.
}
\section{Methods (by class)}{
\itemize{
\item \code{readsData}: Alters the value of the replicates slot of
a readsData object.
}}

\examples{
RD <- new("readsData", counts = ERCC)
reps <- list(A = c(2,4,5,3,10),B = c(9,7,1,8,6))
RD <- setReplicates(RD, reps)
RD@replicates

}
