% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\name{setReplicates}
\alias{setReplicates}
\title{Generic  for the replicates slot of the readsData object.}
\usage{
setReplicates(object, repsval)
}
\arguments{
\item{object}{a readsData object}

\item{repsval}{A list  of column indices for the samples in each condition.}
}
\description{
Generic  for the replicates slot of the readsData object.
}
\keyword{internal}
