% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_denoiSeq.R
\name{setStepSizes}
\alias{setStepSizes}
\title{Generic  for the stepSizes slot of the readsData object.}
\usage{
setStepSizes(object, stepSizesval)
}
\arguments{
\item{object}{a readsData object}

\item{stepSizesval}{A list of step sizes for all the  parameters.}
}
\description{
Generic  for the stepSizes slot of the readsData object.
}
\keyword{internal}
