% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{tunedStepSize}
\alias{tunedStepSize}
\title{Get  values of the tuned step sizes.}
\usage{
tunedStepSize(RDobject)
}
\arguments{
\item{RDobject}{A readsData object with a filled output slot.}
}
\value{
A list of the tuned step sizes of all the parameters.
}
\description{
Extracts the tuned step sizes for each parameter from  the return value of  denoiseq.
}
\examples{
#pre -filtering to remove lowly expressed genes
ERCC <- ERCC[rowSums(ERCC)>20,]
RD <- new('readsData',counts = ERCC)
steps <- 100
#100 steps are not adequate. Just for illustration here.
BI <- denoiseq(RD,steps)
tunedStepSize(BI)

}
